make clear structure
